migration
added Language reference
added model
added migration
change usings
moved models and entities for folders
