--- BooksStore/Model/Entities/AuthorBook.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace Model.Entities
-{
-    public class AuthorBook
-    {
-        public Guid BookId { get; set; }
-        public Book Book { get; set; }
-
-        public Guid AuthorId { get; set; }
-        public Author Author { get; set; }
-    }
-}
