+++ BooksStore/Model/Entities/References/PainterStyle.cs
@@ -1,0 +1,19 @@
+﻿using System;
+
+namespace Model.Entities.References
+{
+    /// <summary>
+    /// Сущность(cправочник) Стиль художника
+    /// </summary>
+    public class PainterStyle
+    {
+        /// <summary>
+        /// Идентификатор
+        /// </summary>
+        public Guid Id { get; set; }
+        /// <summary>
+        /// Наименование
+        /// </summary>
+        public string Name { get; set; }
+    }
+}
