--- BooksStore/ViewModel/Models/AuthorBookModel.cs
@@ -1,13 +1,0 @@
-﻿using System;
-
-namespace ViewModel.Models
-{
-    public class AuthorBookModel
-    {
-        public Guid BookId { get; set; }
-        public BookModel Book { get; set; }
-
-        public Guid AuthorId { get; set; }
-        public AuthorModel Author { get; set; }
-    }
-}
